--- samples/BoltOn.Overrides/Mediator/IQueryUncommitted.cs
@@ -1,9 +1,0 @@
-﻿using BoltOn.Requestor.Interceptors;
-using BoltOn.Requestor.Pipeline;
-
-namespace BoltOn.Overrides.Mediator
-{
-    public interface IQueryUncommitted<out TResponse> : IRequest<TResponse>, IEnableInterceptor<UnitOfWorkInterceptor>
-    {
-    }
-}
