+++ samples/BoltOn.Overrides/Requestor/IQueryUncommitted.cs
@@ -1,0 +1,10 @@
+﻿using BoltOn.Requestor.Interceptors;
+using BoltOn.Requestor.Pipeline;
+
+namespace BoltOn.Overrides.Requestor
+{
+    public interface IQueryUncommitted<out TResponse> : IRequest<TResponse>,
+		IEnableInterceptor<UnitOfWorkInterceptor>
+    {
+    }
+}
